+++ src/tools.hpp
@@ -5,3 +5,5 @@
 #include "file_write.hpp"  // IWYU pragma: export
 #include "shell.hpp"       // IWYU pragma: export
 #include "spawn.hpp"       // IWYU pragma: export
+#include "web_extract.hpp" // IWYU pragma: export
+#include "web_search.hpp"  // IWYU pragma: export